// common/enums/controller-name.enum.ts
export enum ControllerNames {
    ENTRY_ROUTE = 'api',
    AUTH = 'auth',
    USERS = 'users',
    ASSETS = 'assets',
    CREW = 'crew',
    ASSET_CREW = 'asset-crew',
    GEO_FENCES = 'geo-fences',
    HOME = 'home',
    GATE_WAY = "gate-way",
    WORK_SCOPE_AREA = "work-scope-areas",
    LAYERS = 'layers',
    CASES = 'cases',
    ENTITY = 'entity',
    CLIENT = 'client',
    OPERATOR = 'operator',
    // DETECTIONS= 'Detections',
    // DETECTION_TYPES= 'DetectionTypes',
    DETECTION_TYPES= 'detection-types',
    DETECTIONS = 'detections',
    SEVERITY = 'severity',
    CASE_STATUS = 'case-status',
    DETECTION_CONFIG = 'detections-config',
    ALERT = 'alerts',
    ANALYTICS = 'analytics'
}